Files have been rearranged, will still tidy things up, delete unnecessary comments etc
