+++ flask-server/models/content_models.py
@@ -1,5 +1,5 @@
 from flask_sqlalchemy import SQLAlchemy
-from user_models import db
+from app import db
 
 # db = SQLAlchemy()
